+++ react/components/ProductPrice/Installments.js
@@ -1,4 +1,5 @@
-import React, { Component } from 'react'
+import React, { Component, Fragment } from 'react'
+import classNames from 'classnames'
 import { isEmpty } from 'ramda'
 import { FormattedMessage } from 'react-intl'
 import PropTypes from 'prop-types'
@@ -7,12 +8,37 @@ import PricePropTypes from './propTypes'
 
 /** Installments component */
 export default class Installments extends Component {
+  static propTypes = {
+    /** Classes to be applied to the root element */
+    className: PropTypes.string,
+    /** Classes to be applied to installment value element */
+    installmentClass: PropTypes.string,
+    /** Classes to be applied to interest rate element */
+    interestRateClass: PropTypes.string,
+    /** Product installments to be displayed */
+    installments: PricePropTypes.installments,
+    /** Pages editor config to display labels */
+    showLabels: PropTypes.bool.isRequired,
+    /** react-intl function to format the prices*/
+    formatNumber: PropTypes.func.isRequired,
+    /** Options to be passe to the formatNumber function*/
+    currencyOptions: PropTypes.shape({
+      style: PropTypes.string.isRequired,
+      currency: PropTypes.string.isRequired,
+      minimumFractionDigits: PropTypes.number.isRequired,
+      maximumFractionDigits: PropTypes.number.isRequired,
+    }).isRequired
+  }
+  
   render() {
     const {
-      installments,
       showLabels,
       formatNumber,
+      installments,
       currencyOptions,
+      className,
+      installmentClass,
+      interestRateClass
     } = this.props
 
     if (!installments || isEmpty(installments)) {
@@ -45,53 +71,35 @@ export default class Installments extends Component {
     const [installmentsElement, installmentPriceElement, timesElement] = [
       installment.NumberOfInstallments,
       formattedInstallmentPrice,
-      <span key="times">&times;</span>,
+      <Fragment>&times;</Fragment>,
     ].map((element, index) => (
-      <span className="vtex-price-installments__value" key={index}>
+      <span className={installmentClass} key={index}>
         {element}
       </span>
     ))
 
     return (
-      <div className="vtex-price-installments__container c-muted-2">
-        <div className="vtex-price-installments dib">
-          {showLabels ? (
-            <FormattedMessage
-              id="pricing.installment-display"
-              values={{
-                installments: installmentsElement,
-                installmentPrice: installmentPriceElement,
-                times: timesElement,
-              }}
-            />
-          ) : (
-            <span>
+      <div className={classNames('vtex-price-installments__container', className)}>
+        {showLabels ? (
+          <FormattedMessage
+            id="pricing.installment-display"
+            values={{
+              installments: installmentsElement,
+              installmentPrice: installmentPriceElement,
+              times: timesElement,
+            }}
+          />
+        ) : (
+            <Fragment>
               {installmentsElement} {timesElement} {installmentPriceElement}
-            </span>
+            </Fragment>
           )}
-          {!installment.InterestRate && (
-            <span className="pl1">
-              <FormattedMessage id="pricing.interest-free" />
-            </span>
-          )}
-        </div>
+        {!installment.InterestRate && (
+          <div className={classNames('vtex-price-installments__interest-rate', interestRateClass)}>
+            <FormattedMessage id="pricing.interest-free" />
+          </div>
+        )}
       </div>
     )
   }
-}
-
-Installments.propTypes = {
-  /** Product installments to be displayed */
-  installments: PricePropTypes.installments,
-  /** Pages editor config to display labels */
-  showLabels: PropTypes.bool.isRequired,
-  /** react-intl function to format the prices*/
-  formatNumber: PropTypes.func.isRequired,
-  /** Options to be passe to the formatNumber function*/
-  currencyOptions: PropTypes.shape({
-    style: PropTypes.string.isRequired,
-    currency: PropTypes.string.isRequired,
-    minimumFractionDigits: PropTypes.number.isRequired,
-    maximumFractionDigits: PropTypes.number.isRequired,
-  }).isRequired,
 }