+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- Remove Header from `vtex.store-components`.
+- Add several props to style the `ProductPrice` component.
 - Add `classes` prop to ProductName and remove `large` prop.
+
+### Removed
+- `Header` from `vtex.store-components`.
 
 ## [2.6.4] - 2018-11-20
 ### Fixed